Refractor code
less call api
change in data class api

+++ app/src/main/java/fr/epf/min1/velib/model/StationDetail.kt
@@ -3,10 +3,10 @@ package fr.epf.min1.velib.model
 
 data class StationDetail(
     val station_id: Double,
-    val is_installed: Double?,
-    val is_renting: Double?,
-    val is_returning: Double?,
-    val numBikesAvailable: Double?,
-    val numDocksAvailable: Double?,
+    val is_installed: Int?,
+    val is_renting: Int?,
+    val is_returning: Int?,
+    val numBikesAvailable: Int?,
+    val numDocksAvailable: Int?,
     //val num_bikes_available_types: Int?,
 )
